Add update regarding fw issue
For the technically inclined:
As of the initial Flutter 2 release, all builds for this would fail. The [kernel_snapshot] would experience an exception with some references to [flutter_tools]. While investigating, the only most recent change I saw was to do with a clean up but it does not look like a breaking sort of change in this case. So it must be something else.
Similar issues:
- https://github.com/flutter/flutter/issues/77293
- https://github.com/flutter/flutter/issues/77317