Avoid cache clearing on full rebuild
Cache before the change would be cleared on each full rebuild. During development, it is done on hot reload (by default, on each save). As a result, it would remove all files from cache, including the selected, and the appropriate message would then show.
This resolves the issue by only running it initially.