Clear cache when accessing `msgPage`
When it comes to having an exception when a file being copied to cache is too big, the file remains but up to the last byte that could fit. It is not removed.
The Android File Picker utility, at least in Android 9, will crash the first few times up-on file selection if storage is too little.
So, this is to clear that out (avoiding the issue) and any other file that should not be in cache anymore when `msgPage` is in use. After all, any use of `msgPage` essentially means that the server must be halted.