Target SDK 33 (Android 13)
`async_zip` was depending on a `0.` release of `ffi`, of which resulted in `file_picker` upgrading due to it depending on a `2.` release of `ffi`.
As a result, `async_zip` was replaced with something a little more up-to-date.
The new `file_picker` release adds a SDK version check before attempting the READ_EXTERNAL_STORAGE permission. This is important for apps targeting SDK 33 and above.