Use try catch instead while attempting to clear cache
It so appears that there is a race condition as the deleteCache function gets called rapidly as a result of the new approach to rebuild the main UI of the application. In this case, it's being called so fast that even if a check is in place for if a listed file (from the scanned directory) exists, by the time it attempts to delete, the file would already be gone.
So.. while not ideal, it looks like catching the specific exception and then doing nothing with it would be the best way to go in this case.. at least without it being too much effort.