Turn off system UI color change animation
With the colours being set on every rebuild now, the status and navigation bars can suddenly change to a different colour for a moment while changing app state.
This change ensures that it immediately sticks with the intended colour. As this avoids the animation aspect, this may also address another issue regarding a logged internal crash on certain devices supposedly tied to the animation aspect.