Avoid shutdown if theme changes during file selection
There is a bug where if the theme changes at all on the Android native files app (via system theme) then it can force the app to rebuild while importing which can then get it into a state where it accesses the import page too early. One of the checks it would run to is if the imported file exist. This fails and so results in the 'file removed' briefly showing up.
While not a proper fix, this change helps avoid running into the import page too early by showing a loading page instead during the import process should this specific case happen. This loading page may briefly appear outside this case but generally it shouldn't.