Use existing page for modified page
When clearing the app cache, it unfortunately results in a 'modify' file system watcher event being fired, then followed by 'remove'. As 'modified' was not checked before, the second event that comes through would be 'remove' and so would show the relevant page. As it is not clear how to be certain that it really was a cache wipe or the app cache being tampered with, it might be best to show a message that covers both events just in case.
Cases where file removals would be correctly classified internally is renaming or directly removing a selected file (such as while in DAM).