Remove Crowdin
Too much friction to work with. Wouldn't even show the newly added strings that were added between strings (not added to the end of the file) after an update or even a fresh reupload. What's worse, is that it also wiped off a few translations.
It's not really worth the effort, so doing away with it.