Only check for file modification if using shared internal storage (#17)
There looks to be a bug or poorly done logic. Either within the watcher, Android, or the kernel itself.
When the files are stored in app cache, when the app is put into the background and left there for a bit (maybe with certain interactions with different apps), a 'modify' event is fired off for the watched file despite not having been modified since stored in cache.
As this does not affect files stored on internal shared storage and is more so suited for that anyway, that is where the check will be solely enabled for.