Kind of adapt to file_picker unique file directory
As of `file_picker` 7.0.1, the package now stores file selections into their own directories that would be named in milliseconds.

The reason for the change, and this was under Android, was that someone was able to select files with the same file name under different directories. I have yet to come across a file manager that can do this.

As this breaks the the post-import process and multi-selection support is a thing, the files will be moved back to where they originally would go. That is, to not over-complicate the process.

There will be a check for if a file already exists under the same name in the selection made and to rename with unique text added into the file name if so. In this case, as things such as `.tar.gz` exist, the additional text would be put in front of the file name. Not ideal, but it sure beats unintentionally messing up the file extension.